% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_balance.R
\name{rain_balance}
\alias{rain_balance}
\title{rain_balance}
\usage{
rain_balance(precipitation, catchment_area, water_demand, tank_capacity,
  tank_initial_volume = 0, loss_factor)
}
\arguments{
\item{precipitation}{Precipitation data in millimeters (vector)}

\item{catchment_area}{Catchment area in square meters}

\item{water_demand}{Water demand in liters (vector of the same length as precipitation)}

\item{tank_capacity}{Maximum tank volume capacity in liters}

\item{tank_initial_volume}{Tank water volume in liters at the begining of the period}

\item{loss_factor}{Fraction (numeric value ranging from 0 to 1) of the rainwater representing the percentage of water that gets wasted in the harvesting process}
}
\description{
Estimates a balance of potential used rainwater, wasted rainwater and stored rainwater based on precipitation data, catchment_area, water_demand, tank_capacity, tank_initial_volume, loss_factor.
}
\examples{
rain_balance(precipitation = c(1, 2, 3),
             catchment_area = 100,
             water_demand = c(100, 0, 100),
             tank_capacity = 450,
             tank_initial_volume = 0,
             loss_factor = 0.75)
}
\keyword{balance}
\keyword{harvesting,}
\keyword{rainwater}
\keyword{water}
